# lib/api-zod/src/generated/types/courseWaitlistEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseWaitlistEntry {
  id: number;
  email: string;
  name?: string | null;
  source: string;
  createdAt: string;
}